change files to another folder, progress in Service changes
