+++ gcba_licitaciones/src/components/BiddingProrogation.vue
@@ -14,22 +14,19 @@
                 <v-icon>mdi-file-document-plus</v-icon>
               </v-btn>
             </template>
-            <span>Nueva Ampliatoria</span>
+            <span>Nueva Prorroga</span>
           </v-tooltip>
         </v-row>
 
         <v-row>
           <v-data-table
             :headers="headers"
-            :items="extensionDataTable()"
+            :items="prorogationDataTable()"
             hide-default-header
             hide-default-footer
             disable-pagination
             class="elevation-1"
           >
-            <template v-slot:item.Budget="{ item }">
-              <span>{{ priceFormater(item.Budget) }}</span>
-            </template>
 
             <template v-slot:item.actions="{ item }">
               <v-btn-toggle>
@@ -55,31 +52,25 @@
                       v-bind="attrs"
                       v-on="on"
                       small
-                      @click="removeExtension(item)"
+                      @click="removeProrogation(item)"
                     >
                       <v-icon>mdi-file-remove</v-icon>
                     </v-btn>
                   </template>
-                  <span>Borrar Extension</span>
+                  <span>Borrar Prorroga</span>
                 </v-tooltip>
               </v-btn-toggle>
             </template>
           </v-data-table>
         </v-row>
 
-      <v-dialog v-model="addWindow" max-width="500">
+        <v-dialog v-model="addWindow" max-width="500">
         <v-card class="mx-auto my-10" max-width="500">
-          <h2 class="d-flex justify-center pt-5">Nueva Ampliatoria</h2>
+          <h2 class="d-flex justify-center pt-5">Nueva Prorroga</h2>
           <v-container class="pa-3">
-            <v-form onSubmit="return false;" @submit="addExtension()">
-              <v-text-field
-                v-model="extensionCode"
-                label="N de Ampliatoria"
-                :counter="64"
-              ></v-text-field>
-
+            <v-form onSubmit="return false;" @submit="addProrogation()">
               <v-menu
-                v-model="menu2"
+                v-model="menu"
                 :close-on-content-click="false"
                 :nudge-right="40"
                 transition="scale-transition"
@@ -88,8 +79,8 @@
               >
                 <template v-slot:activator="{ on, attrs }">
                   <v-text-field
-                    v-model="extensionDate"
-                    label="Fecha Ampliatoria"
+                    v-model="prorogationDate"
+                    label="Fecha Prorroga"
                     prepend-icon="mdi-calendar"
                     readonly
                     v-bind="attrs"
@@ -97,16 +88,11 @@
                   ></v-text-field>
                 </template>
                 <v-date-picker
-                  v-model="extensionDate"
-                  @input="menu2 = false"
+                  v-model="prorogationDate"
+                  @input="menu = false"
                   locale="es-AR"
                 ></v-date-picker>
               </v-menu>
-
-              <v-currency-field
-                label="Monto Ampliatoria"
-                v-model="budget"
-              ></v-currency-field>
 
               <v-row class="mt-8 mx-auto">
                 <v-btn color="success" class="pa-2" type="submit"
@@ -120,17 +106,11 @@
 
       <v-dialog v-model="editWindow" max-width="500">
         <v-card class="mx-auto my-10" max-width="500">
-          <h2 class="d-flex justify-center pt-5">Editar Ampliatoria</h2>
+          <h2 class="d-flex justify-center pt-5">Editar Prorroga</h2>
           <v-container class="pa-3">
-            <v-form onSubmit="return false;" @submit="editExtencion()">
-              <v-text-field
-                v-model="extensionCode"
-                label="N de Ampliatoria"
-                :counter="64"
-              ></v-text-field>
-
+            <v-form onSubmit="return false;" @submit="editProrogation()">
               <v-menu
-                v-model="menu"
+                v-model="menu2"
                 :close-on-content-click="false"
                 :nudge-right="40"
                 transition="scale-transition"
@@ -139,8 +119,8 @@
               >
                 <template v-slot:activator="{ on, attrs }">
                   <v-text-field
-                    v-model="extensionDate"
-                    label="Fecha Ampliatoria"
+                    v-model="prorogationDate"
+                    label="Fecha Prorroga"
                     prepend-icon="mdi-calendar"
                     readonly
                     v-bind="attrs"
@@ -148,46 +128,40 @@
                   ></v-text-field>
                 </template>
                 <v-date-picker
-                  v-model="extensionDate"
+                  v-model="prorogationDate"
                   @input="menu = false"
                   locale="es-AR"
                 ></v-date-picker>
               </v-menu>
 
-              <v-currency-field
-                label="Monto Ampliatoria"
-                v-model="budget"
-              ></v-currency-field>
-
               <v-row class="mt-8 mx-auto">
-                <v-btn color="success" class="pa-2" type="submit">Editar</v-btn>
+                <v-btn color="success" class="pa-2" type="submit"
+                  >Editar</v-btn
+                >
               </v-row>
             </v-form>
           </v-container>
         </v-card>
       </v-dialog>
+
     </v-container>
   </div>
 </template>
 
 <script>
 export default {
-  data() {
+    data() {
     return {
       headers: [
         //{ text: "id", value: "_id" },
-        { text: "N Ampliatoria", value: "ExtensionCode" },
-        { text: "Fecha Ampliatoria", value: "ExtensionDate" },
-        { text: "Monto", value: "Budget" },
+        { text: "Fecha Prorroga", value: "ProrogationExpired" },
         { text: "Accion", value: "actions", sortable: false },
       ],
       addWindow: false,
       editWindow: false,
       menu: false,
       menu2: false,
-      extensionDate: "",
-      extensionCode: "",
-      budget: 0,
+      prorogationDate: "",
       editIndex: -1,
     };
   },
@@ -195,7 +169,8 @@ export default {
   props: {
     biddingService: Object,
   },
-  methods: {
+
+  methods:{
     showAdd() {
       this.addWindow = !this.addWindow;
     },
@@ -204,58 +179,50 @@ export default {
       this.editWindow = !this.editWindow;
     },
 
-    addExtension() {
-      this.biddingService.ExtensionData.push({
-        ExtensionCode: this.extensionCode,
-        ExtensionDate: this.extensionDate,
-        Budget: this.budget,
-      });
+    addProrogation() {
+      this.biddingService.ProrogationExpired.push(this.prorogationDate);
 
-      this.cleanNewContractorView();
+      this.cleanView();
       this.showAdd();
     },
 
-    cleanNewContractorView() {
-      (this.extensionCode = ""), (this.extensionDate = ""), (this.budget = 0);
+    cleanView() {
+      this.prorogationDate = ''
     },
 
-    extensionDataTable() {
+    prorogationDataTable(){
       //por que no se actulizan los datos si no, cuando se editan los campos de
-      //ExtensionData no se actualiza la vista de tabla
-      let length = this.biddingService.ExtensionData.length;
-      return this.biddingService.ExtensionData.slice(0, length);
+      //ProrogationExpired no se actualiza la vista de tabla
+      let data = []
+      this.biddingService.ProrogationExpired.forEach(date => {
+        data.push({
+            ProrogationExpired: date
+        })
+      })
+      return data
     },
 
     loadEdit(item) {
       this.showEdit();
-      (this.extensionCode = item.ExtensionCode),
-        (this.extensionDate = item.ExtensionDate),
-        (this.budget = item.Budget);
-      this.editIndex = this.biddingService.ExtensionData.indexOf(item);
+      (this.prorogationDate = item.ProrogationExpired),
+      this.editIndex = this.biddingService.ProrogationExpired.indexOf(item.ProrogationExpired);
     },
 
-    removeExtension(item) {
-      let indexEdit = this.biddingService.ExtensionData.indexOf(item);
-      this.biddingService.ExtensionData.splice(indexEdit, 1);
+    removeProrogation(item) {
+      let indexEdit = this.biddingService.ProrogationExpired.indexOf(item);
+      this.biddingService.ProrogationExpired.splice(indexEdit, 1);
     },
 
-    editExtencion() {
-      this.biddingService.ExtensionData[this.editIndex] = {
-        ExtensionCode: this.extensionCode,
-        ExtensionDate: this.extensionDate,
-        Budget: this.budget,
-      };
+    editProrogation() {
+      this.biddingService.ProrogationExpired[this.editIndex] = this.prorogationDate
       this.showEdit();
-      this.cleanNewContractorView();
-    },
-
-    priceFormater(num) {
-      const formatter = new Intl.NumberFormat("es-AR");
-      return formatter.format(num);
+      this.cleanView();
     },
   },
-  computed: {},
-};
+
+}
 </script>
 
-<style></style>
+<style>
+
+</style>
