+++ gcba_licitaciones/src/components/PliegoInputService.vue
@@ -151,6 +151,7 @@
             transition="scale-transition"
             offset-y
             min-width="auto"
+            :disabled="biddingNull"
           >
             <template v-slot:activator="{ on, attrs }">
               <v-text-field
@@ -178,6 +179,7 @@
             transition="scale-transition"
             offset-y
             min-width="auto"
+            :disabled="biddingNull"
           >
             <template v-slot:activator="{ on, attrs }">
               <v-text-field
@@ -205,6 +207,7 @@
             transition="scale-transition"
             offset-y
             min-width="auto"
+            :disabled="biddingNull"
           >
             <template v-slot:activator="{ on, attrs }">
               <v-text-field
@@ -234,6 +237,7 @@
             transition="scale-transition"
             offset-y
             min-width="auto"
+            :disabled="biddingNull"
           >
             <template v-slot:activator="{ on, attrs }">
               <v-text-field
@@ -261,6 +265,7 @@
             transition="scale-transition"
             offset-y
             min-width="auto"
+            :disabled="biddingNull"
           >
             <template v-slot:activator="{ on, attrs }">
               <v-text-field
@@ -288,6 +293,7 @@
             transition="scale-transition"
             offset-y
             min-width="auto"
+            :disabled="biddingNull"
           >
             <template v-slot:activator="{ on, attrs }">
               <v-text-field
@@ -315,6 +321,7 @@
             transition="scale-transition"
             offset-y
             min-width="auto"
+            :disabled="biddingNull"
           >
             <template v-slot:activator="{ on, attrs }">
               <v-text-field
@@ -352,6 +359,7 @@
             v-model="biddingService.PreAdjudgmentActNumber"
             label="N acta Preadjudicada"
             :counter="64"
+            :disabled="biddingNull"
           ></v-text-field>
         </v-col>
       </v-row>
@@ -375,6 +383,7 @@
             transition="scale-transition"
             offset-y
             min-width="auto"
+            :disabled="biddingNull"
           >
             <template v-slot:activator="{ on, attrs }">
               <v-text-field
@@ -402,6 +411,7 @@
             transition="scale-transition"
             offset-y
             min-width="auto"
+            :disabled="biddingNull"
           >
             <template v-slot:activator="{ on, attrs }">
               <v-text-field
@@ -426,6 +436,7 @@
             v-model="biddingService.ApproveNumber"
             label="N de Aprobatoria"
             :counter="64"
+            :disabled="biddingNull"
           ></v-text-field>
         </v-col>
 
@@ -437,6 +448,7 @@
             transition="scale-transition"
             offset-y
             min-width="auto"
+            :disabled="biddingNull"
           >
             <template v-slot:activator="{ on, attrs }">
               <v-text-field
@@ -462,6 +474,7 @@
           <v-currency-field
             label="Monto Adjudicado"
             v-model="biddingService.AllocatedBudget"
+            :disabled="biddingNull"
           ></v-currency-field>
         </v-col>
 
@@ -472,7 +485,9 @@
             readonly
           ></v-text-field>
         </v-col>
+      </v-row>
 
+      <v-row>
         <v-col cols="12" md="3">
           <v-menu
             v-model="menu11"
@@ -481,11 +496,12 @@
             transition="scale-transition"
             offset-y
             min-width="auto"
+            :disabled="biddingNull"
           >
             <template v-slot:activator="{ on, attrs }">
               <v-text-field
                 v-model="biddingService.ContractDate"
-                label="Fecha Perfeccion Orden de Compra"
+                label="Perfeccion Orden de Compra"
                 prepend-icon="mdi-calendar"
                 readonly
                 v-bind="attrs"
@@ -505,11 +521,11 @@
             v-model="biddingService.PurchaseOrder"
             label="Orden de Compra"
             :counter="64"
+            :disabled="biddingNull"
           ></v-text-field>
         </v-col>
-      </v-row>
 
-      <v-col cols="12" md="3">
+        <v-col cols="12" md="3">
           <v-menu
             v-model="menu15"
             :close-on-content-click="false"
@@ -517,6 +533,7 @@
             transition="scale-transition"
             offset-y
             min-width="auto"
+            :disabled="biddingNull"
           >
             <template v-slot:activator="{ on, attrs }">
               <v-text-field
@@ -536,68 +553,44 @@
           </v-menu>
         </v-col>
 
-      <v-row>
         <v-col cols="12" md="3">
           <v-text-field
-            :value="dayQuantity"
-            label="Dias de tramites"
+            :value="totalPOday"
+            label="Dias antes de vencimiento OC"
             :counter="64"
+            readonly
+          ></v-text-field>
+        </v-col>
+
+        <v-col cols="12" md="3">
+          <v-select
+            v-model="biddingService.PurchaseOrderMode"
+            :items= "['ABIERTA', 'CERRADA']"
+            label="Modalidad Orden"
+            required
+            :disabled="biddingNull"
+          ></v-select>
+        </v-col>
+
+        <v-col cols="12" md="3">
+          <v-text-field
+            :value="totalBiddingDays"
+            label="Dias de Tramites Total"
+            :counter="64"
+            readonly
           ></v-text-field>
         </v-col>
       </v-row>
-        
-        <v-col cols="12">
+
+      <v-row>
+        <v-col cols="8">
           <h3 class="justify-center">Ampliatoria</h3>
+          <BiddingExtension :biddingService="biddingService" />
         </v-col>
 
-        <v-col cols="12">
-          <BiddingExtension :biddingService="biddingService"/>
-        </v-col>
-
-        <v-col cols="12" md="3">
-          <v-checkbox
-            v-model="biddingService.Extension"
-            label="Ampliatoria"
-          ></v-checkbox>
-        </v-col>
-
-        <v-col cols="12" md="3">
-          <v-checkbox
-            v-model="biddingService.Prorogation"
-            label="Prorroga"
-          ></v-checkbox>
-        </v-col>
-
-      <v-row v-if="biddingService.Prorogation">
-        <v-col cols="12">
+        <v-col cols="4">
           <h3 class="justify-center">Prorroga</h3>
-        </v-col>
-
-        <v-col cols="12" md="3">
-          <v-menu
-            v-model="menu14"
-            :close-on-content-click="false"
-            :nudge-right="40"
-            transition="scale-transition"
-            offset-y
-            min-width="auto"
-          >
-            <template v-slot:activator="{ on, attrs }">
-              <v-text-field
-                v-model="biddingService.ProrogationExpired"
-                label="Fecha vencimiento prorroga"
-                prepend-icon="mdi-calendar"
-                readonly
-                v-bind="attrs"
-                v-on="on"
-              ></v-text-field>
-            </template>
-            <v-date-picker
-              v-model="biddingService.ProrogationExpired"
-              @input="menu14 = false"
-              locale="es-AR"
-            ></v-date-picker>
-          </v-menu>
+          <BiddingProrogation :biddingService="biddingService" />
         </v-col>
       </v-row>
 
@@ -609,6 +602,7 @@
             label="Observaciones"
             auto-grow
             :counter="300"
+            :disabled="biddingNull"
           ></v-textarea>
         </v-col>
       </v-row>
@@ -622,12 +616,14 @@
 import { mapState, mapMutations } from "vuex";
 import ContractorOffer from "./ContractorOffer.vue";
 import BiddingExtension from "./BiddingExtension.vue";
+import BiddingProrogation from "./BiddingProrogation.vue";
 import Days from "../Helpers/Days";
 
 export default {
   components: {
     ContractorOffer,
-    BiddingExtension
+    BiddingExtension,
+    BiddingProrogation,
   },
   props: {
     showBackBtn: Boolean,
@@ -660,20 +656,32 @@ export default {
   },
   computed: {
     ...mapState({
-      biddingService: state => state.biddingService.biddingService,
-      types: state => state.biddingService.types,
-      status: state => state.biddingService.status,
+      biddingService: (state) => state.biddingService.biddingService,
+      types: (state) => state.biddingService.types,
+      status: (state) => state.biddingService.status,
       directContractType: (state) => state.biddingService.directContract,
-      contractor: state => state.biddingService.contractor,
-      holidays: state => state.biddingService.holidays,
+      contractor: (state) => state.biddingService.contractor,
+      holidays: (state) => state.biddingService.holidays,
     }),
 
     porcentageSPO() {
       return Math.fround(
-        (this.biddingService.AllocatedBudget / this.biddingService.OfficialBudget -1 ) * 100).toFixed(2);
+        (this.biddingService.AllocatedBudget /
+          this.biddingService.OfficialBudget -
+          1) *
+          100
+      ).toFixed(2);
     },
 
-    dayQuantity() {
+    totalBiddingDays() {
+      return Days.daysBetween(
+        this.biddingService.DocumentEntryDate,
+        this.biddingService.ContractDate,
+        this.holidays
+      );
+    },
+
+    totalPOday() {
       return Days.daysBetween(
         this.biddingService.ContractDate,
         this.biddingService.DueDatePO,
@@ -694,9 +702,14 @@ export default {
         ? date.toLocaleDateString()
         : "";
     },
+
+    biddingNull(){
+      return this.biddingService.Status === 'DESIERTA' ? true : false
+    }
+
   },
   methods: {
-    ...mapMutations('biddingService',["cleanPliego"]),
+    ...mapMutations("biddingService", ["cleanPliego"]),
     close() {
       this.$emit("close");
       this.cleanPliego();
