+++ gcba_licitaciones/src/components/statistic/BiddingBudget.vue
@@ -5,20 +5,23 @@
         <v-row>
           <v-col cols="12" md="3">
             <v-select
-              v-model="pickedContractor"
-              :items="onlyNameContractor()"
-              label="Contratista"
+              v-model="budget"
+              :items="budgets"
+              label="Tipo de Presupuesto"
             ></v-select>
           </v-col>
-          </v-row>
+        </v-row>
 
-          <v-row>
-          <v-col>
+        <v-row>
+          <v-col v-if="budget==='Presupuesto Oficial'">
+            Fecha de Llamado entre:
+          </v-col>
+          <v-col v-if="budget==='Monto Adjudicado'">
             Fecha de Contrato entre: 
           </v-col>
         </v-row>  
 
-          <v-row>
+        <v-row>
           <v-col cols="12" md="3">
             <v-menu
               v-model="menu"
@@ -46,7 +49,7 @@
               ></v-date-picker>
             </v-menu>
           </v-col>
-  
+
           <v-col cols="12" md="3">
             <v-menu
               v-model="menu2"
@@ -59,7 +62,7 @@
               <template v-slot:activator="{ on, attrs }">
                 <v-text-field
                   v-model="finishDate"
-                  label="Fecha de fin"
+                  label="Fecha de Fin"
                   prepend-icon="mdi-calendar"
                   readonly
                   v-bind="attrs"
@@ -79,7 +82,20 @@
             <v-checkbox v-model="checkbox" label="Historico"></v-checkbox>
           </v-col>
         </v-row>
-  
+        
+
+        <v-row>
+            <v-col cols="12">
+                Presupuesto Entre:
+            </v-col>
+            <v-col cols="12" md="3">
+              <v-currency-field v-model="botBudget" label="Minimo"></v-currency-field>  
+            </v-col>
+            <v-col cols="12" md="3">
+              <v-currency-field v-model="topBudget" label="Maximo"></v-currency-field>  
+            </v-col>
+        </v-row>
+
 
         <v-row>
           <v-col>
@@ -142,75 +158,82 @@
 </template>
 
 <script>
-import PliegoTable from "../components/PliegoTable.vue";
-import Componentevista from "../components/componentevista.vue";
-import PliegoInput from "../components/PliegoInput.vue";
-import { mapActions, mapState } from "vuex";
+import PliegoTable from "../PliegoTable.vue";
+import Componentevista from "../componentevista.vue";
+import PliegoInput from "../PliegoInput.vue";
+import { mapActions, mapState} from "vuex";
 
 export default {
-    data() {
-        return {
-            status: "",
-            startDate: "",
-            finishDate: "",
-            menu: false,
-            menu2: false,
-            checkbox: false,
-            onePliego: {OfficialBudget: null, AllocatedBudget: null},
-            pliegos:[],
-            viewList: true,
-            viewEdit: false,
-            viewEvery: false,
-            idDelete: "",
-            dialog: false,
-            pickedContractor:"",
-            pliegoHeaders: [
-                { text: "Nº Licitación", value: "BiddingNumber" },
-                { text: "Expediente", value: "Record", width: "130px" },
-                { text: "Expediente BAC OBRAS", value: "RecordBAC" },
-                { text: "Obra", value: "Bidding", width: "300px" },
-                { text: "Repartición", value: "Responsable" },
-                { text: "Responsable", value: "Division" },
-                { text: "Tipo Licitación", value: "BiddingType" },
-                { text: "Presupuesto Oficial", value: "OfficialBudget" },
-                { text: "Estado", value: "Status" },
-                // {text:"Revisión Pliegos ingreso", value: "EntryDocumentReview"},
-                // {text:"Revisión Pliegos egreso", value: "ExitDocumentReview"},
-                // {text:"1ª Salida a PG", value: "FirstPG"},
-                // {text:"1ª Vuelta de PG", value: "FirstLapPG"},
-                { text: "Fecha Llamado", value: "CallDate", width: "200px" },
-                { text: "Fecha Apertura de Ofertas", value: "BidOpeningDate" },
-                // {text:"Cantidad de Ofertas", value: "BidQuantity"},
-                // {text:"Fecha Acta Preadjudic.", value: "PreAdjudgmentActDate"},
-                // {text:"Nº Acta Preadj.", value: "PreAdjudgmentActNumber"},
-                // {text:"2º Salida a PG", value: "SecondPG"},
-                // {text:"2ª vuelta de PG", value: "SecondLapPG"},
-                // {text:"Cant. Días", value: "DayQuantity"},
-                // {text:"N° Aprobatoria", value: "ApproveNumber"},
-                // {text:"Fecha Aprobatoria", value: "ApproveDate"},
-                // {text:"Monto Adjudicado", value: "AllocatedBudget"},
-                // {text:"% S/P.O.", value: "SPO"},
-                // {text:"Contratista", value: "Contractor"},
-                {text:"Fecha Contrato", value: "ContractDate"},
-                // {text:"Dias de trámite", value: "ProcedureDays"},
-                // {text:"OBSERVACIONES", value: "Observations"},
-                { text: "Accion", value: "actions", sortable: false },
-            ],
-        }
-    },
-    components:{
-        PliegoTable,
-        PliegoInput,
-        Componentevista,
-    },
-    methods: {
+  data() {
+    return {
+      status: "",
+      startDate: "",
+      finishDate: "",
+      botBudget: 0,
+      topBudget: 0,
+      menu: false,
+      menu2: false,
+      checkbox: false,
+      budgets: [
+        "Presupuesto Oficial",
+        "Monto Adjudicado",
+      ],
+      onePliego: {OfficialBudget: null, AllocatedBudget: null},
+      pliegos: [],
+      viewList: true,
+      viewEdit: false,
+      viewEvery: false,
+      idDelete: "",
+      dialog: false,
+      budget: "",
+      viewAll: true,
+      pliegoHeaders: [
+        { text: "Nº Licitación", value: "BiddingNumber" },
+        { text: "Expediente", value: "Record", width: "130px" },
+        { text: "Expediente BAC OBRAS", value: "RecordBAC" },
+        { text: "Obra", value: "Bidding", width: "300px" },
+        { text: "Repartición", value: "Responsable" },
+        { text: "Responsable", value: "Division" },
+        { text: "Tipo Licitación", value: "BiddingType" },
+        { text: "Presupuesto Oficial", value: "OfficialBudget" },
+        { text:"Monto Adjudicado", value: "AllocatedBudget"},
+        { text: "Estado", value: "Status" },
+        // {text:"Revisión Pliegos ingreso", value: "EntryDocumentReview"},
+        // {text:"Revisión Pliegos egreso", value: "ExitDocumentReview"},
+        // {text:"1ª Salida a PG", value: "FirstPG"},
+        // {text:"1ª Vuelta de PG", value: "FirstLapPG"},
+        { text: "Fecha Llamado", value: "CallDate" },
+        { text: "Fecha Apertura de Ofertas", value: "BidOpeningDate" },
+        // {text:"Cantidad de Ofertas", value: "BidQuantity"},
+        // {text:"Fecha Acta Preadjudic.", value: "PreAdjudgmentActDate"},
+        // {text:"Nº Acta Preadj.", value: "PreAdjudgmentActNumber"},
+        // {text:"2º Salida a PG", value: "SecondPG"},
+        // {text:"2ª vuelta de PG", value: "SecondLapPG"},
+        // {text:"Cant. Días", value: "DayQuantity"},
+        // {text:"N° Aprobatoria", value: "ApproveNumber"},
+        // {text:"Fecha Aprobatoria", value: "ApproveDate"},
+        // {text:"% S/P.O.", value: "SPO"},
+        // {text:"Contratista", value: "Contractor"},
+        { text: "Fecha Contrato", value: "ContractDate" },
+        // {text:"Dias de trámite", value: "ProcedureDays"},
+        // {text:"OBSERVACIONES", value: "Observations"},
+        { text: "Accion", value: "actions", sortable: false },
+      ],
+    };
+  },
+  components: {
+    PliegoTable,
+    PliegoInput,
+    Componentevista,
+  },
+  methods: {
     ...mapActions('bidding',[
-      'biddingContractor',
+      "biddingBudget",
       "loadEditPliego",
       "editPliego",
       "deletePliego"
     ]),
-    
+
     viewOne(pliego) {
       this.onePliego = pliego;
       this.viewList = false;
@@ -240,48 +263,31 @@ export default {
       this.viewEdit = false;
     },
 
-    onlyNameContractor(){
-      let contractorName = []
-      this.contractor.forEach(element => contractorName.push(element.Name))
-      return contractorName
+    async validate() {
+      let query = {};
+      query.budget = this.budget === "Presupuesto Oficial" ? "OfficialBudget" : "AllocatedBudget";
+      query.botBudget = this.botBudget
+      query.topBudget = this.topBudget
+
+      if (this.checkbox) {
+        query.startDate = new Date(1000, 1, 1);
+        query.finishDate = new Date(3000, 1, 1);
+      } else {
+        query.startDate = this.startDate;
+        query.finishDate = this.finishDate;
+      }
+
+      this.pliegos = await this.biddingBudget(query);
     },
-
-    contractorNametoId(contractorName){
-      let idContractor = "";
-      this.contractor.forEach((contractor) =>
-        contractor.Name === contractorName
-          ? (idContractor = contractor._id)
-          : 0
-      );
-      return idContractor;
-    },
-
-    async validate(){
-        let query = {}
-        query.contractor = this.contractorNametoId(this.pickedContractor)
-        if (this.checkbox) {
-                query.startDate = new Date(1000, 1, 1)
-                query.finishDate = new Date(3000, 1, 1)
-        }
-        else {
-            query.startDate = this.startDate
-            query.finishDate = this.finishDate
-        }
-
-        this.pliegos = await this.biddingContractor(query)
-    }  
-          
   },
+  
   computed: {
     ...mapState({
       contractor: state => state.bidding.contractor,
       holidays: (state) => state.bidding.holidays,
     })
-  },
-}
-
+  }
+};
 </script>
 
-<style>
-
-</style>
+<style></style>
