+++ gcba_licitaciones/src/components/statistic/charts/BarChart.vue
