+++ gcba_licitaciones/src/components/statistic/charts/DoughnutChart.vue
