+++ gcba_licitaciones/src/store/modules/biddingService.js
@@ -42,11 +42,8 @@ export default {
       SPO: 0,
       Contractor: null,
       ContractDate: "",
-      Extension: false,
       ExtensionData: [],
-      Prorogation: false,
-      ProrogationExpired: "",
-      ProcedureDays: "",
+      ProrogationExpired: [],
       Observations: "",
     },
     pliegos: [],
@@ -90,11 +87,8 @@ export default {
         SPO: 0,
         Contractor: null,
         ContractDate: "",
-        Extension: false,
         ExtensionData: [],
-        Prorogation: false,
-        ProrogationExpired: "",
-        ProcedureDays: "",
+        ProrogationExpired: [],
         Observations: "",
       };
     },
